+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.2.0] - 2020-07-11
 ### Added
 - Method `fallback()`
 - Equality operator support
@@ -61,7 +62,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial version.
 
-[Unreleased]: https://github.com/f3ath/maybe-just-nothing/compare/0.1.9...HEAD
+[Unreleased]: https://github.com/f3ath/maybe-just-nothing/compare/0.2.0...HEAD
+[0.2.0]: https://github.com/f3ath/maybe-just-nothing/compare/0.1.9...0.2.0
 [0.1.9]: https://github.com/f3ath/maybe-just-nothing/compare/0.1.8...0.1.9
 [0.1.8]: https://github.com/f3ath/maybe-just-nothing/compare/0.1.7..0.1.8
 [0.1.7]: https://github.com/f3ath/maybe-just-nothing/compare/0.1.6..0.1.7